Patched bug and added more security to tree state checks
Patched a bug where the program would try to split a pixel in two because the actual node would have a resolution of 2x1 and a color not similar enough with lower color difference tresholds. Added if(branch) check in tree state checks and an else which prints an error and terminates the program.